requirements.txt
scikit-surprise==1.1.4
pandas==2.2.3
numpy==1.26.4